NXCM-4454: Description was wrong for almost all actions.
The "description" might be passed along (from config or CLI) by use
to send it to Nexus, and it will be stored (and emailed) along with
Staging actions. The default value (when user did not pass in any)
was wrong, as it was referring always to "closed by nexus-staging-maven-plugin".

Now, in default case (user does not provide it), the description is
vastly improved: not only it tells the proper action, but also contains
Peter's bonus too.