NXCM-5194: G level MD deploy
As turns out, those are locally staged, but not remotely
deployed into staging repo (or to hosted repo in case of
deferred deploy, both are affected).

Reason was that "remote deploy" is driven by an "index file"
that simply enlists all the artifacts deployed, but NOT
metadata. This is a problem in case of G level MD, as
other MDs are handled transparently by deployer, while
G level MD is not.

This is merely a "quickfix" for the problem, it makes
it work, but has it's own compromises.