NEXUS-6538: Proxy protocol is already decided
Proxy was queried for nexusUrl already, and we got
ONE proxy to be used.

Still, in cases when user has a HTTP proxy defined
(and selected by Maven) to connect to HTTPS nexus,
this would cause "ProxyInfo and BaseUrl protocols
does not align!" error.

Reason is simple: we asked for and got a Proxy to
be used with _nexusUrl_'s protocol, and NOT
the protocol Proxy had set in settings.xml

TODO: We should inspect logic in
org.sonatype.maven.mojo.settings.MavenSettings#selectServer
as it does "fallback" to HTTP proxy if no HTTPS
proxy set for HTTPS remote. Still, Maven settings
reference is _unclear_ can a user define multiple
or only one active Proxy. As per doco, it say

  "active: true if this proxy is active.
  This is useful for declaring a set of proxies,
  but only one may be active at a time."

Hence, having multiple _active_ Proxy entries
in settings might be considered as illegal even.